DitherPunk

Documentation for DitherPunk.

DitherPunk.bayer

DitherPunk.bayer_dithering

DitherPunk.linear2srgb

DitherPunk.ordered_dithering

DitherPunk.print_braille

DitherPunk.show_dither

DitherPunk.srgb2linear

DitherPunk.bayer — Method

Contruct (un-normalized) Bayer matrices through recursive definition.

DitherPunk.bayer_dithering — Method

Ordered dithering using the Bayer matrix as a threshold map.

DitherPunk.linear2srgb — Method

Convert from linear to sRGB color space.

DitherPunk.ordered_dithering — Method

Ordered dithering using threshold maps. https://en.wikipedia.org/wiki/Ordereddithering#Pre-calculatedthreshold_maps

DitherPunk.print_braille — Method

(Ab)use spy to interpret image as sparse array and print "sparsity pattern" in braille through UnicodePlots.

DitherPunk.show_dither — Method

Apply Gray to indicate that image should be interpreted as a grayscale image.

DitherPunk.srgb2linear — Method

Convert from sRGB to linear color space.