API Reference

DitherPunk.BinaryDitherMetric

DitherPunk.ClosestColor

DitherPunk.ConstantThreshold

DitherPunk.OrderedDither

DitherPunk.SeparateSpace

DitherPunk.WhiteNoiseThreshold

DitherPunk.Atkinson

DitherPunk.BalancedCenteredPoint

DitherPunk.Bayer

DitherPunk.Burkes

DitherPunk.CentralWhitePoint

DitherPunk.ClusteredDots

DitherPunk.FalseFloydSteinberg

DitherPunk.Fan93

DitherPunk.FloydSteinberg

DitherPunk.JarvisJudice

DitherPunk.Rhombus

DitherPunk.ShiauFan

DitherPunk.ShiauFan2

DitherPunk.Sierra

DitherPunk.SierraLite

DitherPunk.Stucki

DitherPunk.TwoRowSierra

DitherPunk.bayer_matrix

DitherPunk.closest_color

DitherPunk.dither

DitherPunk.dither!

DitherPunk.gradient_image

DitherPunk.linear2srgb

DitherPunk.srgb2linear

DitherPunk.test_on_gradient

DitherPunk.upscale

Functions

DitherPunk.BinaryDitherMetric — Type

Define custom color difference metric with linear distances between grayscale values.

DitherPunk.ClosestColor — Type

Simplest form of image quantization by turning each pixel to the closest one in the provided color palette cs. Technically this not a dithering algorithm as the quatization error is not "randomized".

DitherPunk.ConstantThreshold — Type

ConstantThreshold(threshold)

Use a constant threshold map. Defaults to 0.5 if threshold isn't specified.

DitherPunk.OrderedDither — Type

OrderedDither(mat::AbstractMatrix)

Generalized ordered dithering algorithm using a threshold map. Takes a normalized threshold matrix mat.

When applying the algorithm to an image, the threshold matrix is repeatedly tiled to match the size of the image. It is then applied as a per-pixel threshold map. Optionally, this final threshold map can be inverted by selecting invert_map=true.

DitherPunk.SeparateSpace — Type

SeparateSpace(alg::AbstractDither)

A meta-algorithm that takes any gray-scale dithering algorithm and applies channel-wise binary dithering.

Note

The output of this algorithm depends on the color type of input image. RGB is recommended.

Examples

julia> dither!(img, SeparateSpace(Bayer()))

DitherPunk.WhiteNoiseThreshold — Type

WhiteNoiseThreshold()

Use white noise as a threshold map.

DitherPunk.Atkinson — Method

Atkinson()

Error diffusion algorithm using the filter

    *   1   1
1   1   1
    1               (1//8)

DitherPunk.BalancedCenteredPoint — Method

BalancedCenteredPoint()

Balanced centered point ordered dithering. Uses $6 \times 6$ threshold matrix BALANCED_CENTERED_POINT_MAT.

DitherPunk.Bayer — Method

Bayer(level)

Ordered dithering using the Bayer matrix as a threshold matrix. The Bayer matrix is of dimension $2^{n+1} \times 2^{n+1}$, where $n$ is the level, which defaults to 1.

[1] Bayer, B.E., "An Optimum Method for Two-Level Rendition of Continuous Tone Pictures," IEEE International Conference on Communications, Conference Records, 1973, pp. 26-11 to 26-15.

DitherPunk.Burkes — Method

Burkes()

Error diffusion algorithm using the filter

        *   8   4
2   4   8   4   2
1   2   4   2   1   (1//42)

[1] Burkes, D., "Presentation of the Burkes error filter for use in preparing continuous-tone images for presentation on bi-level devices." Unpublished, 1988.

DitherPunk.CentralWhitePoint — Method

CentralWhitePoint()

Central white point ordered dithering. Uses $6 \times 6$ threshold matrix CENTRAL_WHITE_POINT_MAT.

DitherPunk.ClusteredDots — Method

ClusteredDots()

Clustered dots ordered dithering. Uses $6 \times 6$ threshold matrix CLUSTERED_DOTS_MAT.

DitherPunk.FalseFloydSteinberg — Method

FalseFloydSteinberg()

Error diffusion algorithm using the filter

*   3
3   2         (1//8)

Occasionally, you will see this filter erroneously called the Floyd-Steinberg filter.

Note

There is no reason to use this algorithm, which is why DitherPunk doesn't export it.

[1] http://www.efg2.com/Lab/Library/ImageProcessing/DHALF.TXT

DitherPunk.Fan93 — Method

Fan93()

Error diffusion algorithm using the filter

      *  7
1  3  5            (1//16)

A modification of the weights used in the Floyd-Steinberg algorithm.

[1] Z. Fan, "A Simple Modification of Error Diffusion Weights", IS&T's 46th Annual Conference, May 9-14, 1993, Final Program and Advanced Printing of Paper Summaries, pp 113-115 (1993).

DitherPunk.FloydSteinberg — Method

FloydSteinberg()

Error diffusion algorithm using the filter

    *   7
3   5   1     (1//16)

[1] Floyd, R.W. and L. Steinberg, "An Adaptive Algorithm for Spatial Gray Scale." SID 1975, International Symposium Digest of Technical Papers, vol 1975m, pp. 36-37.

DitherPunk.JarvisJudice — Method

JarvisJudice()

Error diffusion algorithm using the filter

        *   7   5
3   5   7   5   3
1   3   5   3   1   (1//48)

Also known as the Jarvis, Judice, and Ninke filter.

[1] Jarvis, J.F., C.N. Judice, and W.H. Ninke, "A Survey of Techniques for the Display of Continuous Tone Pictures on Bi-Level Displays," Computer Graphics and Image Processing, vol. 5, pp. 13-40, 1976.

DitherPunk.Rhombus — Method

Rhombus()

Diagonal ordered matrix with balanced centered points. Uses $8 \times 8$ threshold matrix RHOMBUS_MAT.

DitherPunk.ShiauFan — Method

ShiauFan()

Error diffusion algorithm using the filter

        *   4
1   1   2           (1//8)

[1] J. N. Shiau and Z. Fan. "Method for quantization gray level pixel data with extended distribution set", US 5353127A, United States Patent and Trademark Office, Oct. 4, 1993

DitherPunk.ShiauFan2 — Method

ShiauFan2()

Error diffusion algorithm using the filter

            *   8
1   1   2   4       (1//16)

[1] J. N. Shiau and Z. Fan. "Method for quantization gray level pixel data with extended distribution set", US 5353127A, United States Patent and Trademark Office, Oct. 4, 1993 [2] J. N. Shiau and Z. Fan. "A set of easily implementable coefficients in error-diffusion with reduced worm artifacts" Color Imaging: Device-Independent Color, Color Hard Copy, and Graphics Arts, volume 2658, pages 222–225. SPIE, March 1996.

DitherPunk.Sierra — Method

Sierra()

Error diffusion algorithm using the filter

        *   5   3
2   4   5   4   2
    2   3   2       (1//32)

Also known as Sierra3 or three-row Sierra due to the filter shape.

DitherPunk.SierraLite — Method

SierraLite()

Error diffusion algorithm using the filter

    *   2
1   1               (1//4)

Also known as Sierra-2-4A filter.

DitherPunk.Stucki — Method

Stucki()

Error diffusion algorithm using the filter

        *   8   4
2   4   8   4   2
1   2   4   2   1   (1//42)

[1] Stucki, P., "MECCA - a multiple-error correcting computation algorithm for bilevel image hardcopy reproduction." Research Report RZ1060, IBM Research Laboratory, Zurich, Switzerland, 1981.

DitherPunk.TwoRowSierra — Method

TwoRowSierra()

Error diffusion algorithm using the filter

        *   4   3
1   2   3   2   1   (1//16)

Also known as Sierra2.

DitherPunk.bayer_matrix — Method

bayer_matrix(n::Int)::AbstractMatrix{Int}

Contruct (un-normalized) Bayer matrices of level n through recursive definition.

DitherPunk.closest_color — Method

Return color in ColorScheme cs that is closest to color according to colordiff

DitherPunk.dither — Function

dither([T::Type,] img, alg::AbstractDither, args...; kwargs...)

Dither img using algorithm alg.

Output

If no return type is specified, the default return type for binary dithering algorithms is Gray{Bool}. For color algorithms, the type of the colorscheme is used.

DitherPunk.dither! — Function

dither!([out,] img, alg::AbstractDither, args...; kwargs...)

Dither img using algorithm alg.

Output

If out is specified, it will be changed in place. Otherwise img will be changed in place.

DitherPunk.gradient_image — Method

Create linear grayscale gradient image of type Gray.

DitherPunk.linear2srgb — Method

Convert from linear to sRGB color space.

DitherPunk.srgb2linear — Method

Convert from sRGB to linear color space.

DitherPunk.test_on_gradient — Method

Test dithering algorithm alg on linear grayscale gradient from gradient_image and show stacked plot.

DitherPunk.upscale — Method

Upscale image by repeating individual pixels.